age_checker_2.0
This Python program collects information about 4 people, including their name, age, and gender. During the input process, it performs the following tasks:

Calculates the average age of all 4 individuals.

Counts how many women are under 20 years old.

Identifies the oldest man and stores his name and age.

After gathering all the information, it prints:

The average age of the group.

The number of women under 20 years old.

The name and age of the oldest man (if any man was entered).